fix(ui): move authOptions to a separate file
